pyproject.toml
[project]
name = "parking-management-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.1",
    "paho-mqtt>=2.1.0",
    "psycopg2>=2.9.11",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.48",
    "supabase>=2.28.0",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_test_loop_scope = "function"
asyncio_default_fixture_loop_scope = "function"
